refactor: make Instructions op a byte

+++ components/compiler-kotlin/app/src/main/kotlin/io/littlelanguages/bendu/compiler/ByteBuilder.kt
@@ -11,7 +11,7 @@ class ByteBuilder {
     }
 
     fun appendInstruction(op: Instructions): ByteBuilder =
-        append(op.op.toByte())
+        append(op.op)
 
     fun appendInt(value: Int): ByteBuilder {
         val bytes = ByteArray(4)
